Rename getLocalizedString and made helper function accept lists of strings.
