+++ components/survey-editor/components/ItemEditor/editor/components/content-editor/SurveyEndEditor.tsx
@@ -4,7 +4,7 @@ import { SurveyContext } from '@/components/survey-editor/surveyContext';
 import { checkMissingTranslations } from '@/components/survey-editor/utils/localeUtils';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
-import { getLocalizedString } from '@/utils/getLocalisedString';
+import { getLocalizedString } from '@/utils/localizedStrings';
 import { generateTitleComponent } from 'case-editor-tools/surveys/utils/simple-generators';
 import React, { useContext, useEffect } from 'react';
 import { LocalizedString, SurveyItem, SurveySingleItem } from 'survey-engine/data_types';
